+++ src/graphql_sqlalchemy/testing.py
@@ -1,8 +1,13 @@
 from __future__ import annotations
 
-from typing import Any
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Any, TypeVar
 
 from graphql import GraphQLEnumType, GraphQLEnumValueMap, GraphQLList, GraphQLNonNull, GraphQLType, is_equal_type
+from sqlalchemy import JSON, Dialect, TypeDecorator
+
+if TYPE_CHECKING:
+    from sqlalchemy.types import TypeEngine
 
 
 def assert_equal_gql_type(a: GraphQLType, b: GraphQLType) -> None:
@@ -23,3 +28,35 @@ def assert_equal_gql_type(a: GraphQLType, b: GraphQLType) -> None:
 
 def mk_comparable_values(values: GraphQLEnumValueMap) -> dict[str, Any]:
     return {k: v.value for k, v in values.items()}
+
+
+T = TypeVar("T")
+
+
+class JsonArray(TypeDecorator[Sequence[T]]):
+    impl = JSON
+    cache_ok = True
+
+    item_type: TypeEngine[T]
+
+    def __init__(self, item_type: TypeEngine[T] | type[TypeEngine[T]], none_as_null: bool = False):
+        super().__init__(none_as_null=none_as_null)
+        self.item_type = item_type() if isinstance(item_type, type) else item_type
+
+    def process_bind_param(self, value: Sequence[T] | None, dialect: Dialect) -> Sequence[T] | None:
+        if value is None:
+            return None
+        if not isinstance(value, Sequence):
+            raise ValueError("value must be a sequence")
+        if not all(isinstance(v, self.item_type.python_type) for v in value):
+            raise ValueError(f"all values must be of type {self.item_type.python_type}")
+        return value
+
+    def process_result_value(self, value: Any | None, dialect: Dialect) -> list[T]:
+        if not isinstance(value, list):
+            raise ValueError("value must be a list")
+        return value
+
+    @property
+    def python_type(self) -> type[list[T]]:
+        return list
